+++ intents/connectors/dialogflow_es/names.py
@@ -1,9 +1,10 @@
 from typing import Type
 
 from intents import Intent
+from intents.helpers.misc import camel_to_snake_case
 
 def context_name(intent_cls: Type[Intent]) -> str:
-    return "c_" + intent_cls.name.replace(".", "_").lower() # TODO: refine
+    return "c_" + camel_to_snake_case(intent_cls.name.replace(".", "_")) # TODO: refine
 
 def event_name(intent_cls: Type[Intent]) -> str:
     """
@@ -12,4 +13,4 @@ def event_name(intent_cls: Type[Intent]) -> str:
     >>> event_name('test.intent_name')
     'E_TEST_INTENT_NAME'
     """
-    return "E_" + intent_cls.name.upper().replace('.', '_')
+    return "E_" + camel_to_snake_case(intent_cls.name.replace(".", "_")).upper()